calculate score

+++ src/GuessThePrice.WebApp/Requests/AddResponse.cs
@@ -2,6 +2,6 @@ using GuessThePrice.Core.Model;
 
 namespace GuessThePrice.WebApp.Requests;
 
-public record PromotionalPriceResponseRequest(decimal Value);
+public record PromotionalPriceResponseRequest(double Value);
 
 public record AddResponseRequest(int ProductId, PromotionalPriceResponseRequest PromotionalPriceResponse);
